abstract  checkExecution
